LUAD.b2.v1

+++ get_dependencies.sh
@@ -11,7 +11,28 @@ exomeBedDir=$6
 exomeBedFile=$7
 
 ## copy BamMaps
-cp ${bamMapDir}${bamMapFile} ${inputDir}${bamMapFile}
-cp ${refDir}${refFile} ${inputDir}${refFile}
-cp ${exomeBedDir}${exomeBedFile} ${inputDir}${exomeBedFile}
- 
+if [ -s ${inputDir}${bamMapFile} ]
+then
+	echo "bamMap is available"
+else
+	echo "bamMap is being copied"
+	cp ${bamMapDir}${bamMapFile} ${inputDir}${bamMapFile}
+fi
+
+## copy reference file
+if [ -s ${inputDir}${refFile} ]
+then
+        echo "refFile is available"
+else
+        echo "refFile is being copied"
+	cp ${refDir}${refFile} ${inputDir}${refFile}
+fi
+
+## copy exome bed file
+if [ -s ${inputDir}${exomeBedFile} ]
+then
+        echo "exome target bed file is available"
+else
+        echo "exome target bed file is being copied"
+	cp ${exomeBedDir}${exomeBedFile} ${inputDir}${exomeBedFile}
+fi
